requirements.txt
numpy~=1.20.2
pygame~=2.0.1
gym~=0.18.0
pandas~=1.2.4
setuptools~=54.0.0
neat-python~=0.92